refactor(camelot): move create test function and types to utils
